requirements.txt
# SmartSec Python Dependencies
flask>=3.0
firebase-admin>=6.0
python-dotenv>=1.0
RPi.GPIO
mfrc522
RPLCD
